pyproject.toml
[tool.poetry]
name = "myshop"
version = "0.1.0"
description = ""
authors = ["hruntov <hruntov@outlook.com>"]

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.6"
flake8 = "^7.1.0"
pillow = "^10.4.0"
celery = "^5.4.0"
psycopg2-binary = "^2.9.9"
flower = "^2.0.1"
requests = "2.31.0"
isort = "^5.13.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
